Changed short to int16_t

+++ LidarController.h
@@ -321,7 +321,7 @@ class LidarController {
         - Read the measured value from data registers
     *******************************************************************************/
     //uint8_t distance(uint8_t Lidar, int * data) {
-	uint8_t distance(uint8_t Lidar, short * data) {
+	uint8_t distance(uint8_t Lidar, int16_t * data) {
       uint8_t distanceArray[2];
       uint8_t nackCatcher = I2C.readWord(lidars[Lidar]->address, MEASURED_VALUE_REGISTER, distanceArray);
       shouldIncrementNack(Lidar, nackCatcher);
@@ -356,7 +356,7 @@ class LidarController {
         This has to be worked on, this is the original implementation without the 
           blocking architecture
     *******************************************************************************/
-    int velocity(uint8_t Lidar, int * data) {
+    int16_t velocity(uint8_t Lidar, int16_t * data) {
       // Set in velocity mode
       I2C.write(lidars[Lidar]->address, REG_ACQ_CONFIG, DATA_VELOCITY_MODE_DATA);
       //  Write 0x04 to register 0x00 to start getting distance readings
@@ -365,7 +365,7 @@ class LidarController {
       uint8_t velocityArray[1];
       uint8_t nack = I2C.readByte(lidars[Lidar]->address, REG_VELOCITY, velocityArray);
 
-      return((int)((char)velocityArray[0]));
+      return((int16_t)((char)velocityArray[0]));
     };
 
     /*******************************************************************************
@@ -405,7 +405,7 @@ class LidarController {
       We could use the async() method in ACQUISITION_DONE, but it would need to spin
       one time more before starting the acquisition again
     *******************************************************************************/
-    uint8_t distanceAndAsync(uint8_t Lidar, short * data) {
+    uint8_t distanceAndAsync(uint8_t Lidar, int16_t * data) {
       uint8_t nackCatcher = distance(Lidar, data);
       // if error reading the value, try ONCE again
       if (nackCatcher)
